Tome of Ghosts -menu update
There's a book icon on the top. It opens a menu which has story and tools tabs. Story tab opens to text view that tells the game's story. Tools opens to Tools tab which describes the items in the game. Pressing/clicking buttons make a soundeffect sound, same as the game menu.